[chore] plan participants schemas
