+++ src/features/plan/participant/servers/schemas/req.schema.ts
@@ -1,0 +1,4 @@
+import { z } from "zod";
+import { extendZodWithOpenApi } from "@asteasolutions/zod-to-openapi";
+
+extendZodWithOpenApi(z);
