+++ src/server/repos/plans/participants/plan-participants.repo.ts
@@ -1,4 +1,5 @@
 import { prismaClient } from "@/server/db/prismaClient";
+import { ParticipantPermissionEnum } from "@/features/plan/participant/type";
 
 const planParticipantsRepo = {
   async insertMany({
@@ -6,7 +7,10 @@ const planParticipantsRepo = {
     participantDataList,
   }: {
     planId: string;
-    participantDataList: { userId: string; permission: string[] }[];
+    participantDataList: {
+      userId: string;
+      permission: ParticipantPermissionEnum[];
+    }[];
   }) {
     return prismaClient.plan_participants.createMany({
       data: participantDataList.map(({ userId, permission }) => ({
@@ -28,7 +32,10 @@ const planParticipantsRepo = {
     participantDataList,
   }: {
     planId: string;
-    participantDataList: { userId: string; permission: string[] }[];
+    participantDataList: {
+      userId: string;
+      permission: ParticipantPermissionEnum[];
+    }[];
   }) {
     const updatePromises = participantDataList.map(({ userId, permission }) =>
       prismaClient.plan_participants.update({
